// JcCar-client/src/app/shared/services/model/actionPlan.ts
/**
 * ALERTA API
 * SMS - Módulo Alertas de SMS que tem como objetivo comunicar aos interessados deliberações das
 * Agências Reguladoras e eventos de segurança relevantes para que as medidas necessárias sejam tomadas
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AlertInvolved } from './alertInvolved';
import { AttachmentActionPlan } from './attachmentActionPlan';


export interface ActionPlan {
  alertInvolved?: AlertInvolved;
  attachmentsForActionPlan?: Array<AttachmentActionPlan>;
  createdBy?: string;
  creationDate?: Date;
  description?: string;
  endDate?: Date;
  id?: number;
  rootCause?: string;
  status?: ActionPlan.StatusActionPlanEnum;
}
export namespace ActionPlan {

  export enum StatusActionPlanEnum  {
      CREATED = 0,
      SENDED = 1,
      IN_APPROVEMENT = 2,
      APPROVED = 3,
      DISAPPROVED = 4
  }
}