// JcCar-client/src/app/shared/services/model/alert.ts
/**
 * ALERTA API
 * SMS - Módulo Alertas de SMS que tem como objetivo comunicar aos interessados deliberações das
 * Agências Reguladoras e eventos de segurança relevantes para que as medidas necessárias sejam tomadas
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AlertClass } from './alertClass';
import { AlertClassification } from './alertClassification';
import { AlertInvolved } from './alertInvolved';
import { Attachment } from './attachment';
import { GoldRole } from './goldRole';
import { Origin } from './origin';
import { StatusApplication } from './statusApplication';
import { Temporality } from './temporality';


export interface Alert {
    actionPlain?: string;
    alertClass?: AlertClass;
    alertClassification?: AlertClassification;
    alertInvolveds?: Array<AlertInvolved>;
    attachments?: Array<Attachment>;
    basicCause?: string;
    createdBy?: string;
    creationDate?: string;
    description?: string;
    goldRoles?: Array<GoldRole>;
    id?: number;
    codigo?: string;
    idOccurrence?: number;
    mainManagement?: string;
    occurrenceDate?: string;
    origin?: Origin;
    sendDate?: string;
    status?: StatusApplication;
    tags?: Array<string>;
    temporality?: Temporality;
    title?: string;
}